Add files via upload
This is a Windows activator that works for LTSC 2019 and 2021. I also used it in one of my videos.

It doesn't work with any other versions of windows. This is for Windows 10 MTR users who don't want to spend money just to use their ocmputers properly.
Legal or not, I don't care. Microsoft already has enough money so this won't affect them in any way.